Revert "[ModuleToKore Refactoring] Creatign `GenerateMapCeilAxioms` compiler pass (#3529)"
This reverts commit e64221b86ad520f5dc9b01ea62a82cb98423e5a2.
